Refactored for TFRecord Compatibility
Refactored code to be compatible with swapping out TFRecords in the data_loader class. This required refactoring all modules to work inception-style and to use the Tensorflow pipelin (except in the loader itself, optionally). This puts the loader in command of building most of the feed_dict, and tracking the image indices to loop over. I think it's ok.

Also a typo fix in my_load_mat.